Refactor socket impl for expansion
