--- app/emitter.ts
@@ -1,4 +1,0 @@
-import cache from '@/utils/global_cache'
-import { EventEmitter } from 'stream'
-
-export default cache('emitter', () => new EventEmitter())
