+++ app/pages/api/socket/index.ts
@@ -1,8 +1,8 @@
 import db from '@/app/db'
-import emitter from '@/app/emitter'
 import { Server } from 'socket.io'
 import { createAdapter } from '@socket.io/postgres-adapter'
 import { PgDatabase } from '@/utils/orm/pg'
+import onSocket from '@/app/extensions/socket'
 
 export default function SocketHandler(req: any, res: any) {
   if (res.socket.server.io) {
@@ -23,22 +23,7 @@ export default function SocketHandler(req: any, res: any) {
       io.adapter(createAdapter((db as PgDatabase).pool))
     }
     res.socket.server.io = io
-    io.on('connection', (client) => {
-      console.debug(`${client.id} connected`)
-      client.on('join', (id) => {
-        console.debug(`${client.id} joined ${id}`)
-        client.join(id)
-        emitter.emit(`join:${id}`)
-        emitter.on(`close:${id}`, () => {
-          console.debug(`${client.id} disconnected from ${id}`)
-          client.leave(id)
-          client.send('close')
-        })
-      })
-      client.on('http:recv', ({ id, ...rest}) => {
-        emitter.emit(`http:recv:${id}`, rest)
-      })
-    })
+    io.on('connection', onSocket)
   }
   res.end()
 }
